Información de estado del dispensador movida del fragment_dispenser.xml al fragment_home.xml
